requirements.txt
pathway[xpack-llm]
python-dotenv
google-generativeai
requests
